+++ Example/AdsExample/AdsExample/AdsExampleApp.swift
@@ -10,15 +10,13 @@ import SwiftUI
 
 @main
 struct AdsExampleApp: App {
-  @StateObject private var adsManager = makeAdsManager()
+  init() {
+    setupDependencies()
+  }
 
   var body: some Scene {
     WindowGroup {
-      ContentView(adsManager: adsManager)
-        .onAppear {
-          adsManager.startGoogleMobileAds()
-          adsManager.preloadConfiguredSlots()
-        }
+      ContentView()
     }
   }
 }