+++ Example/AdsExample/AdsExample/InterstitialDemoView.swift
@@ -4,10 +4,11 @@
 //
 
 import AdsKit
+import SwiftInjected
 import SwiftUI
 
 struct InterstitialDemoView: View {
-  @ObservedObject var adsManager: AdsKitManager
+  @InjectedObservable var adsManager: AdsKitManager
 
   @State private var statusMessage = "Tap \"Load\" to load an interstitial ad."
   @State private var isLoaded = false